fix: unblock process because when input is not read
Rationale: on some occasions a process might just block if the STDOUT is not read. In that case as with the non Thread variant we open the error stream and we don't get data, hence block forever. With this change it is ensure that both streams are getting read from.